Refactor import/export code into seperate classes
